requirements.txt
pytest==8.2.2
pytest-bdd==7.2.0
requests==2.32.3
jsonschema==4.23.0
python-dotenv==1.0.1
pytest-html==4.1.1
allure-pytest==2.13.5
pytest-xdist==3.5.0
urllib3==2.5.0